backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==3.0.1
google-generativeai==0.3.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
reportlab==4.0.9
pandas==2.1.4
bleach==6.1.0
email-validator==2.1.0
python-dateutil==2.8.2
